chore: move locale-flag to root level of components folder
